Basic entity added
Entity will eventually be a linked list, and the game will own the list,
looping through it for updates and rendering. Not sure if I'll make the
player entity a generic entity like everything else yet...

For now, the player is just roughed in during the main loop. Basic event
handling added as well, so the player sprite zips around the screen with
the arrow keys - fun!