Working on Entity
Continued working on Entity object. The WSL_App object now owns a
(double) linked list of entities. Why is it a double linked list? I
don't know, but it seemed ideal and was literally like three extra lines
of code. Entity creation still needs some love - right now the player is
created before the main loop and added to the entity list; the player
should be created during a "new_game()" routine. Entity destruction
seems to work well - added a "projectile" entity that is created when
the user presses the space bar and destroyed when it exists the screen.

Entities have flags that (when combined) say what they are - player,
enemy, player projectile, etc - and if they are alive/dead. At the end
of the update routine the game's entity list is ran through and "dead"
entities are destroyed.

Entities have a "cooldown" timer as a basic integer, and the idea is
that they can only "act" if the cooldown flag is off. Different actions
can set different cooldown times, turn on the cooldown flag, and wait
for the update loop to turn it off. While update is cleaning up dead
entities, if an entities cooldown timer is greater than 0, it reduces it
by 1. If the timer is 0, it turns off the cooldown flag.

The main loop has an SDL_Delay added at the end to keep the refresh at
60fps - I think. I stole the idea from looking at Parallel Realities SDL
tutorials, and he doesn't explain why he put that there. It seems to
work well enough.