Updated keypress handling, other minor refactoring
Moved entity update functions to update.h/c - because it made sense.
Space was lonely so I added in some asteroids.. and then brought in some
RNG action to spice up the asteroids. Then decided the asteroids needed
to spin towards the player...

WSL_App is getting kinda bulky - turning into a catchall structure and I
don't really like it. I might just leave it like that though since the
goal of this "game" isn't really to make a game, it's just to play
around with SDL/C.

The Parallel Reality SDL tutorials do a neat trick with the keyboard
event handling, and I took that idea and gave it my own flair.