Player can now take damage and die!
Needs some more work, but the player ship can take damage. When the
player takes a hit, the ship flashes and the player is invulnerable for
a short time (and can't fire lasers). Rendering the ship "damage"
sprites is in progress, as well as a "heads up"/hull integrity meter in
the UI (Well, the whole UI doesn't actually exist anywhere outside my
imagination... yet.)