Fonts and text on screen!
Added in SDL_ttf functions to load a font, and a function to create a
SDL_Texture with a formatted text string (variadics). It doesn't seem
efficient to create and destroy textures every frame - but it doesn't
seem to matter? It works well for now, anyways. Added a simple HUD
display with the player's health (hull integrity) and points at the top.
(Player can't score points yet)