UFOs!
Started working on some UI stuff, got bored, and decided to play with
UFOs. UFOs in this game are spawned randomly, outside of the normal
"waves" of enemies (that don't exist yet). The trick here was making the
UFO movement seem "natural" and not predictable.

Added in the Vec2i files from the toolbox (a bit late) and created
Vec2f. Vec2f is used to make a bezier curve from a couple points on the
screen (start/middle/end), and then the t value of the linear
interpolation is what the UFO advances each update cycle to "move"
across the screen. This makes the UFO move in a smooth curve from left
to right across the screen. Heckin neat. To spice it up, the UFO
sometimes randomly changes the start/end point when it reaches the
end/start point - so it looks like the path curve dynamically changes.

Right now, UFOs spawn randomly (10% chance) when the player shoots down
a (big) asteroid.

For the UI stuff - Kenney has some awesome ui graphics that I would love
to implement... but haven't figured out how I want them to look - yet.
Right now I have the full spritesheet, but it might make more sense to
make static graphics and then my own spritesheet for the UI stuff. I
haven't decided if that would be better than dynamically breaking apart
the graphics in code and then reassembling them larger/smaller as needed.