UFOs now annoyingly fire lasers at the player
UFOs fire lasers - a bit slower than the player. Started with having
them fire the lasers AT the player but that was absurd. Lasers are fired
straight down - enemy lasers can't hit anything but the player, which is
neat. I might adjust that to make it so that UFOs can shoot down
asteroids. Lasers can't hit other lasers.

Also, now when a UFO takes damage it speeds up while it's invulnerable -
zipping around the screen.