Generic "Timer" added in Entity
Exploring an idea I had - what if there was a simple linked list of integer
"timers" that each entity owns. Each timer could have a void
(*action)(void *owner) function that is called when the timer is updated, and a
boolean "repeat" if the timer resets when it hits 0? This could also be
used elsewhere in the application (say for enemy waves, asteroid
spawning, etc)...

Added this idea in timer.h/c, implemented as a double linked list.