Basic state framework added
Reviving this project after a hiatus - Realized this is almost 80% of a
"complete" game... just missing a menu screen, and proper new game/game
over states. Added in the basic ideas for the states, along with
Update/Events/Draw functions for each state (that are called by the
regular update/handle events/draw function). The player can hold
spacebar at the menu to make absurd fireworks.