Added fancy flashing text entity
For showing alerts/popups/whatever. Super quick and easy to add in - the
text is an entity that is created at a location and during it's update
routine it just adjusts its alpha +- 100... giving it a cool space text
vibe. Right now just used for a "good luck" message at the beginning of
the game, it may be used later for powerups, when the player dies (Be
careful! Ships are expensive!), or something else (WARNING!).