Fixed minor bug in wsl_texture_load_text
Boy that function is a mouthful. I should probably write some better
string/text handling functions... ANYWAY - this function was not using
variadic arguments correctly, it wasn't checking for the size of the
formatted string before allocating memory for the string to render,
causing the string to truncate. Easy fix.