Pickups!
Added a "pickup" entity that moves straight down the screen and can only
contact the player... and if they do contact the player they do
something!

Interestingly, this are basically the same thing as projectiles and I
probably could have used a lot of the projectile code instead of writing
new stuff... but this makes things more clear. For now, UFOs spawn a
"shield" pickup that gives the player back a point of shield strength.
Interestingly (maybe), 2 points was wildly overpowered. More pickups to
come!